Unified all CX program makers, getters, adders, etc. into a single file (ast.go).
